requirements/base.txt
Django==1.8.*

# can go back to a tagged release when https://github.com/django-admin-bootstrapped/django-admin-bootstrapped/pull/235
# included in the tagged release
#django-admin-bootstrapped>2.5.7
-e git+https://github.com/django-admin-bootstrapped/django-admin-bootstrapped.git@master#egg=django_admin_bootstrapped

# Custom storages for files - django-storages 1.6.6 and above do not support django 1.8
boto3
django-storages<1.6.6

django-apptemplates>=1.2

django-authtools>1.4.0

django-bootstrap3>=6.2.0

#django-compat>=1.0

django-ckeditor>=5.0.2

django-compressor>=1.6

django-filebrowser-no-grappelli>=3.6.1

django-filter>=0.13.0,<1.0

Django-floppyforms>=1.6.0

django-form-utils>=1.0

django-hijack>=2.1

django-import-export>=0.2

django-permanent>=1.0.12

django-select2>=5.5

#django-suit>=0.2

django-tables2>=1.2.5

-e git+https://github.com/levic/django-webpack-loader.git@webpack_loader_class_0_3_0#egg=django-webpack-loader

djangorestframework>=3.2

#future>=0.15

lxml>=3.6.4

mandrill>=1.0.6

MySQL-python>=1.2

Pillow>=3.0

#psycopg2>=2.5

python-dateutil>=2.5.3

pytz

rules>=1.1.1

spyne>=2.12.11

# only used for CI
snakefood

#six>=1.9

suds>=0.4

tzlocal==1.2

unicodecsv>=0.14.1

Unipath>=1.1

xlrd>=0.9.4

xmltodict>=0.10.2

# ------------------------------------------------
# CI Dependencies

# for reset_db, runserver_plus
# runserver_plus is only used on CI, not live
django-extensions>=1.7.6
werkzeug
freezegun

sqlparse>=0.2.3
